refactor: 클래스 및 패키지 정리
