+++ src/main/java/racingcar/input/carname/validator/exception/DuplicateCarNamesException.java
@@ -1,4 +1,4 @@
-package racingcar.input.utils.carname.validators.exceptions;
+package racingcar.input.carname.validator.exception;
 
 public class DuplicateCarNamesException extends RuntimeException {
     public DuplicateCarNamesException(String message) {
