+++ src/main/java/racingcar/input/carname/validator/exception/InvalidCarNameFormatException.java
@@ -1,4 +1,4 @@
-package racingcar.input.utils.carname.validators.exceptions;
+package racingcar.input.carname.validator.exception;
 
 public class InvalidCarNameFormatException extends RuntimeException {
     public InvalidCarNameFormatException(String message) {
