+++ src/test/java/racingcar/input/racingtrytime/RacingTryTimeValidatorUtilsTest.java
@@ -1,4 +1,4 @@
-package racingcar.input.utils.racingtrytime;
+package racingcar.input.racingtrytime;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.input.utils.racingtrytime.validators.RacingTryTimeValidatorUtils;
-import racingcar.input.utils.racingtrytime.validators.exceptions.NotNaturalNumberRacingTimeException;
+import racingcar.input.racingtrytime.validator.RacingTryTimeValidatorUtils;
+import racingcar.input.racingtrytime.validator.exception.NotNaturalNumberRacingTimeException;
 
 public class RacingTryTimeValidatorUtilsTest {
